move groupAllAreaSeries to separate method
